src/requirements.txt
drf_yasg2
asgiref==3.4.1
Django==2.2.10
django-rest-framework==0.1.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
simplejwt==2.0.1
djoser